refactored decrypt and encrypt
